Revert "feat(sqlite): add WASM conditional import isolation for logd_sqlite"
This reverts commit 8f0abd83bcadd46d46481e0424739eab63d74eb5.
